be fixed

+++ models/chatModel.js
@@ -29,6 +29,7 @@ exports.updateChatSessionTitle = async (request, session_id) => {
 };
 
 exports.getChatMessageBySessionId = async (session_id, user_id) => {
+  console.log("user_id:", user_id);
   return await pool.query(
     "SELECT  c.request,c.response FROM  chats c JOIN chat_sessions s ON c.session_id=s.chat_session_id where c.session_id =$1 AND s.user_id=$2 ORDER BY c.created_at",
     [session_id, user_id]
@@ -36,19 +37,33 @@ exports.getChatMessageBySessionId = async (session_id, user_id) => {
 };
 //added to filter chats by user logged in
 exports.getSessionCheckByUserId = async (sessionID, user_id) => {
+  console.log("user_id:", user_id);
+
   return await pool.query(
     "Select * FROM chat_sessions WHERE chat_session_id=$1 AND user_id=$2",
     [sessionID, user_id]
   );
 };
-exports.deleteChatBySessionId = async (session_id) => {
+exports.deleteChatBySessionId = async (session_id, user_id) => {
+  console.log(
+    "sessionid to delete :",
+    session_id,
+    "/n user_id to delete:",
+    user_id
+  );
   return await pool.query(
-    "DELETE  from chat_sessions where chat_session_id=$1",
-    [session_id]
+    "DELETE  from chat_sessions where chat_session_id=$1 and user_id=$2",
+    [session_id, user_id]
   );
 };
 exports.getAnyExistingSessionId = async () => {
   return await pool.query(
     "SELECT session_id from chats where session_id IS NOT NULL LIMIT 1 "
+  );
+};
+exports.createChatSessionInDB = async (user_id) => {
+  return await pool.query(
+    "INSERT INTO chat_sessions (user_id) VALUES ($1) RETURNING chat_session_id",
+    [user_id]
   );
 };
